reafactor code base, implement SOLID principle
